W4 more database tweaks, 3dmap good enough, route plotter good
